# .futurxlab/inknowing-api-spec.yaml
openapi: 3.0.3
info:
  title: InKnowing API
  description: |
    知应(InKnowing) AI知识对话平台 MVP API规范
    基于业务逻辑守恒原理设计，确保与用户旅程、时序图、状态图完全对应
  version: 1.0.0
  contact:
    name: FuturX Lab
    email: support@inknowing.com
servers:
  - url: http://localhost:3000/api
    description: 本地开发环境
  - url: https://api.inknowing.com
    description: 生产环境
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: 用户登录
      description: |
        使用邮箱密码登录，创建24小时session
        状态转换: GUEST → LOGGING_IN → LOGGED_IN
        引用: #REF-LOGIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  session:
                    type: string
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /books/{id}:
    get:
      tags: [Books]
      summary: 获取书籍详情
      description: |
        获取单本书籍的完整信息
        状态转换: BROWSING_BOOKS → VIEWING_DETAIL
        引用: #REF-DETAIL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 书籍详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetail'
  /books/{id}/characters:
    get:
      tags: [Books, Characters]
      summary: 获取书籍角色列表
      description: |
        获取书中可对话的角色
        状态转换: VIEWING_DETAIL → LOADING_CHARACTERS → BOOK_READY
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 角色列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  characters:
                    type: array
                    items:
                      $ref: '#/components/schemas/Character'

  # ===== 对话相关 API (CONVERSATION_STATE_MACHINE) =====
...
  /admin/login:
    post:
      tags: [Admin]
      summary: 管理员登录
      description: |
        使用环境变量配置的密码登录管理后台
        状态转换: ADMIN_GUEST → ADMIN_AUTHENTICATING → ADMIN_LOGGED_IN
        引用: #REF-ADMIN-LOGIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin_session:
                    type: string
...
  /admin/books:
    post:
      tags: [Admin]
      summary: 创建书籍
      description: |
        基于识别结果创建新书籍
        状态转换: BOOK_IDENTIFIED → BOOK_CREATING → BOOK_CREATED
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
...
  /admin/characters/{id}:
    put:
      tags: [Admin]
      summary: 编辑角色信息
      description: |
        更新角色的所有配置
        状态转换: CHARACTER_EDITING → CHARACTER_SAVED
        引用: #REF-CHAR-MANAGE
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharacterUpdateRequest'
      responses:
        '200':
          description: 更新成功
  /admin/config/ai:
    get:
      tags: [Admin]
      summary: 获取AI配置
      description: |
        获取当前AI服务配置
        状态: AI_CONFIG
      security:
        - adminAuth: []
      responses:
        '200':
          description: AI配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIConfig'

    put:
      tags: [Admin]
      summary: 更新AI配置
      description: |
        配置通义千问和其他AI服务
        状态转换: CONFIG_EDITING → CONFIG_TESTING → CONFIG_SAVED
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIConfig'
      responses:
        '200':
          description: 配置更新成功
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
...
    Book:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        cover_url:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        ai_score:
          type: integer
        status:
          type: string
          enum: [online, offline]
...
    BookDetail:
      allOf:
        - $ref: '#/components/schemas/Book'
        - type: object
          properties:
            description:
              type: string
            publisher:
              type: string
            publish_date:
              type: string
            conversation_strategy:
              type: string
              enum: [ai_native, rag_only, hybrid]
            has_document:
              type: boolean
            character_count:
              type: integer
    BookInfo:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        description:
          type: string
        publisher:
          type: string
        publish_date:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
    BookCreateRequest:
      allOf:
        - $ref: '#/components/schemas/BookInfo'
        - type: object
          required: [ai_score]
          properties:
            ai_score:
              type: integer
            cover_url:
              type: string
...
    Character:
      type: object
      properties:
        id:
          type: string
        book_id:
          type: string
        name:
          type: string
        description:
          type: string
        personality:
          type: string
        speaking_style:
          type: string
        background:
          type: string
        prompt_template:
          type: string
...
    CharacterUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        personality:
          type: string
        speaking_style:
          type: string
        background:
          type: string
        prompt_template:
          type: string
...
    AIConfig:
      type: object
      properties:
        qwen_api_key:
          type: string
        qwen_model:
          type: string
          default: qwen-max
        embedding_model:
          type: string
        openai_compatible:
          type: object
          properties:
            base_url:
              type: string
            api_key:
              type: string
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: 用户Session认证

    adminAuth:
      type: apiKey
      in: cookie
      name: admin_session
      description: 管理员Session认证

# 业务逻辑守恒验证标记
# 所有API端点都与以下视图完全对应：
# 1. User Journey: 用户旅程中的每个动作都有对应API
# 2. Sequence Diagram: 时序图中的每个交互都有API支撑
# 3. State Diagram: 状态转换的每个触发点都是API调用
# 4. 双向可推导: 从API可以还原完整业务流程